requirements.txt
pyTelegramBotAPI
qrcode
pillow
pyzbar